requirements.txt
Flask==3.0.0
Flask-Cors==4.0.1
Pillow==10.4.0
python-dotenv==1.0.1
pydicom==2.4.4
numpy==1.26.4
tensorflow==2.16.1
google-generativeai==0.7.1
